chore: add a package that computes `s_curve` profile
The method is based on `Trajectory Planning for Automatic Machines and Robots, ch 3.4.4`. And
I've run some samples in the book to manually verify the profile. However, there might be some
issues in the code (as it always have bug...), so I'll try to fix these potential bugs when I merge the
`s_curve` to nucleo firmware.