chore: apply new packet decode function in rx task and fix proto encoding error in tx task
1. The packet decode function is updated to return valid packet index, so rx task is updated accordingly
2. In tx task:
- The `join` is used to wait until all the data are set
- Move `stream` and `pb_encoder` to loop. The reason is that the `pb_encoder` will keep pushing new
data to `stream`, so it will eventually fail to encode the packet. Since the `pb_encoder` doesn't provide
a method to clean its internal buffer, I have to put the data inside to loop to avoid this encoding issue.

Note, I'm not sure the performance impact when `stream` and `pb_encoder` are moved to the loop.
In my naive thoughts, the `stream` and `pb_encoder` are fixed size, so it shouldn't impact performance
that much, but I might be wrong on that point, and I might need to change it in the future if it really hurt
the performance.