chore: update velocity error threshold to `60.0` in `ready()`
Based on current settings, the motor velocity(rpm) is calculated as:
`encoder_count_diff * 30.0`

Since encoder is counting the `counts`, the minimum `count_diff` is 1, so the minimum
threshold is 30.0 (the motor can't detect the velocity under 30 rpm). And to avoid other
potential errors(Ex: encoder lost count), so I set the threshold to 60.0.

Note: this value might be updated in the future since it is based on encoder count, so I might
tune it in the future.