chore: run `motion_task` with `InterruptExecutor` and use `Watch` to pass motion data
## Motivation on moving `motion_task` to `InterruptExecutor`

In previous commit, the motor was vibrating when I sent velocity command. Because all the tasks
are running on the same executors, the `motion_task` might be influenced by other tasks, Ex:
(I could be highly wrong, so it is just a rough mental picture on why the motor was vibrating)
1. timer interrupt get triggered -> signal `TIMER_SIGNAL`
2. executor decided to run other tasks
3. after other tasks are finished, run `motion_task`
and this might caused the `motion_task` not running periodically and made the motor vibrate.

To solve this issue, I decide to try `multiprio` example in `embassy` which is a method to run the task
with different priority. After some trials, the `motion_task` is moved to `TIM5` interrupt executor which
has higher priority than other tasks. After testing it with several velocity commands, look like it is working

Note:
The another method that I came up with was to move the `motion_task` directly into timer interrupt, but
this required `static mut` to let timer interrupt access motion data which seemed a little bit dirty to me(better
methods might exist, and I might look for them when I have other similar issues), so I decided to try `multiprio`.

---

## Motivation on using `Watch` to pass motion data

I think the latest motion data is more valuable to me, so it doesn't require the `queue`, therefore I use `Watch`
to pass latest motion data.