chore: add pub functions to return current encoder count, actual position, and update velocity calculation function
## Motivation on adding public functions to return current encoder count, actual position

The `pub` is removed from `curr_enc_count`, so a `get` function is added to return encoder count.
And for actual position function, it calculates actual position in `rad`(basically the same with the one
that returns encoder count), and it will be used in the ROS2 hardware interface to provide position
feedback for ROS2 diff-drive controller.