chore: spawn threads to perform serial communication
The updates are mainly on `start`. `thread::scope`is used in previous commit, but
`thread::scope` is guaranteed to join all the threads at the end of scope, so it can't
provide the effect I need. So, this parts are updated to use normal `thread::spawn`.

Since `thread::spawn` can't use `self`, I used  associated function in closure. Also,
the sender/receiver in struct is updated to `Option` because these members need
to be initialized in `start`