chore: remove time-delay in thread tasks
I think the sleep delay in thread tasks is not good to control the data (we might
need to tune manully for proper dealy), so I decide to remove them.

For `tx_task`, it will fire when it receives a data from channel, so I used `recv()` to
block the thread. And for `rx_task`, I let the `read` do its work, since it might block
internally to read the data.