chore!: create a queue to receive user commands
In previous commit, channel is used as communication method. But if
`Stop` command is used, it would be hard to purge all the commands in
channel, so I decide to switch to queue + mutex

The `CommandRxQueue` is used to store user commands, and it will notify
`tx_task` if a new command is pushed into it. Also, if it encounters
`Stop` command, then all the commands in the queue will be purged, so we
can handle `Stop` command first.

Also, the previous version used a atomic variable to terminate `tx_task`
and `rx_task`. And now:
1. `tx_task`, it gets terminated if it receives a `Unspecified` command
2. `rx_task`, it gets terminated if atomic variable is set to true