refactor: create different size for pub/sub channel and internal motion command queue
The `CHANNEL_SIZE` is used in `PubSubChannel` and `MOTION_CMD_QUEUE_SIZE` is used
in motion struct. If the queue in motion struct is full, I want to make sure there are spaces in
`PubSubChannel`, so `Halt` command can be sent to motion struct. And for the other
commands, since they don't have the same priority as `Halt`, the sender needs to wait
until there are spaces in the queue.