refactor: improve command readiness check in run() method
In previous version, I used control mode to check if the command can be set
or not. But this method is not correct, because it blocks `Halt` command.
To fix this, I will first check the command in queue, and if the command can
be set, then it will be popped from queue.