refactor: separate the async task on receiving command and sending command
In previous version, the receiver and sender work together in 1 async task. But
this caused issues when the buffer in target board is full, and the commands
aren't sent properly.

Ex: the UI sends 6 commands to the async task, and it encounters buffer full on
5th command. The async task sets all the commands in once (1ms per command,
I measured from board), the 5th and 6th commands will never be sent to the board,
 because the command channel between UI and async task is empty.

So I fix this part by separating receiver and sender, and the sender will try to
re-send the command if the buffer in target is full.