fix: update direction in `stop` method to ensure velocity is properly decelerated to 0
I found this issue when running position commands that have negative displacement. When commands are
stopped in the middle, the velocity will jump to 0 which makes motor rotate drastically.

This issue comes from the incorrect direction settings in `stop` method. Before the fix, it uses 0.0 as displacement
and makes `set_target` misjudge the direction that causes incorrect sign in later calculation.

To fix the issue, I change the displacement to `self.target_data.dir` which is the direction of current segment to
make sure `set_target` set correct direction. And a test case is added to for this fix.

Note:
the displament setting in `stop` is only used to update direction and it will be updated to 0 later in `set_target` to
deceleration immediately.